requirements.txt
pandas
# requests
# matplotlib
# seaborn
# plotly
jupyter
# langchain
langchain-community
faiss-cpu